pyproject.toml
[project]
name = "bayes-ii"
version = "0.1.0"
description = ""
authors = [
    {name = "ToledoSnows",email = "felipe.toledo.neves@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "pymc (>=5.26.1,<6.0.0)",
    "streamlit (>=1.51.0,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "plotly (>=6.3.1,<7.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "scipy (>=1.16.3,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
